Fixed up tests for notifications.js
