+++ lib/constants.js
@@ -82,7 +82,7 @@ export const NOTIFICATION = {
     CONNECT_REPO: "Connect your repo with CodeSync",
     CONNECT_AFTER_JOIN: "Successfully logged in to CodeSync. Let's connect your repo",
     CHOOSE_ACCOUNT: "Choose account to sync your repo",
-    USE_DIFFERENT_ACCOUNT: "Use differnet account",
+    USE_DIFFERENT_ACCOUNT: "Use differet account",
     PUBLIC_OR_PRIVATE: "Do you want to make the repo public?",
     REPO_SYNCED: "Repo synced successfully! Track it at",
     BRANCH_SYNCED: "Branch synced successfully! Track it at",
