+++ lib/init/init_handler.js
@@ -15,6 +15,7 @@ import {checkServerDown, getUserForToken} from "../utils/api_utils";
 import {initUtils} from './utils';
 import {askAndTriggerSignUp} from '../utils/auth_utils';
 import {generateSettings} from "../settings";
+import {askPublicPrivate} from "../utils/notifications";
 
 
 export const syncRepo = async (repoPath, accessToken, viaDaemon = false, isSyncingBranch = false) => {
@@ -120,23 +121,8 @@ const postSyncignoreUpdated = async (repoPath, repoName, branch, user, accessTok
         return;
     }
 
-    const notification = atom.notifications.addInfo(
-        NOTIFICATION.PUBLIC_OR_PRIVATE, {
-            buttons: [
-                {
-                    text: NOTIFICATION.YES,
-                    onDidClick: () => postClickVisibility(repoPath, repoName, branch, user, accessToken,
-						true, viaDaemon, isRepoSynced, notification)
-                },
-                {
-                    text: NOTIFICATION.NO,
-                    onDidClick: () => postClickVisibility(repoPath, repoName, branch, user, accessToken,
-						false, viaDaemon, isRepoSynced, notification)
-                },
+    askPublicPrivate();
 
-            ],
-            dismissable: true
-        });
 };
