+++ lib/utils/notifications.js
@@ -3,6 +3,7 @@ import { NOTIFICATION, USER_PATH } from '../constants';
 import { syncRepo } from '../init/init_handler';
 import { readYML } from './common';
 import { logout, redirectToBrowser } from './auth_utils';
+import {generateSettings} from "../settings";
 
 
 export const showSignUpButtons = () => {
@@ -66,7 +67,8 @@ export const showConnectRepo = (repoPath, email="", accessToken="") => {
 
 export const showChooseAccount = (repoPath) => {
     // Check if access token is present against users
-    const users = readYML(USER_PATH);
+    const settings = generateSettings();
+    const users = readYML(settings.USER_PATH);
     const accounts = [];
     Object.keys(users).forEach(key => {
         const user = users[key];
@@ -82,6 +84,7 @@ export const showChooseAccount = (repoPath) => {
 
 	const emails = accounts.map(account => account.email);
 	const options = [...emails, NOTIFICATION.USE_DIFFERENT_ACCOUNT];
+
     const notification = atom.notifications.addInfo(
 		NOTIFICATION.CHOOSE_ACCOUNT, {
             buttons: options.map((option) => {
@@ -112,4 +115,24 @@ const postSelectionChooseAccount = async (repoPath, accounts, selection) => {
     // We have token, repoPath Trigger Init
     await syncRepo(repoPath, user.access_token);
 };
+
+export const askPublicPrivate = () => {
+    const notification = atom.notifications.addInfo(
+        NOTIFICATION.PUBLIC_OR_PRIVATE, {
+            buttons: [
+                {
+                    text: NOTIFICATION.YES,
+                    onDidClick: () => postClickVisibility(repoPath, repoName, branch, user, accessToken,
+                        true, viaDaemon, isRepoSynced, notification)
+                },
+                {
+                    text: NOTIFICATION.NO,
+                    onDidClick: () => postClickVisibility(repoPath, repoName, branch, user, accessToken,
+                        false, viaDaemon, isRepoSynced, notification)
+                },
+
+            ],
+            dismissable: true
+        });
+}
